Migrate Choropleth to React: skeleton
