+++ client/app/visualizations/choropleth/maps/index.js
@@ -1,0 +1,7 @@
+import countriesDataUrl from './countries.geo.json';
+import subdivJapanDataUrl from './japan.prefectures.geo.json';
+
+export {
+  countriesDataUrl,
+  subdivJapanDataUrl,
+};
